prod_deep_agent_main/requirements.txt
deepagents==0.4.11
langgraph==1.1.2
langchain==1.2.12
langgraph-checkpoint-postgres==3.0.4
psycopg[binary]==3.3.3
python-dotenv==1.2.2
typer==0.24.1
rich==14.3.3
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langchain_mcp_adapters
